add earn rewards modal for cosmos
